main/requirements.txt
telethon==1.40
python-dotenv==1.0.0
